requirements.txt
anyio==3.6.1
autopep8==1.7.0
certifi==2022.6.15
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
pycodestyle==2.9.1
python-dotenv==0.20.0
rfc3986==1.5.0
sniffio==1.2.0
tekore==4.4.0
toml==0.10.2
